vdpimporthelper/requirements.txt
asgiref==3.6.0
black==22.12.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
Django==4.1.5
django-jazzmin==2.6.0
flake8==6.0.0
google-api-core==2.11.0
google-api-python-client==2.75.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.8.0
googleapis-common-protos==1.58.0
httplib2==0.21.0
idna==3.4
mccabe==0.7.0
mypy-extensions==0.4.3
numpy==1.24.2
oauthlib==3.2.2
pandas==1.5.3
pathspec==0.10.3
platformdirs==2.6.2
protobuf==4.21.12
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pyflakes==3.0.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.14
xmltodict==0.13.0
